# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kwaaka Integration API",
        "title": "Integration API",
        "contact": {
            "name": "Kwaaka",
            "email": "neversi123123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "staging.kwaaka.com",
    "basePath": "/v1",
    "paths": {
...
        "/v1/order": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "external"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.CreationResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    }
                }
            }
        },
        "/v1/order/{order_id}": {
            "get": {
                "tags": [
                    "external"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order_id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "external"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order_id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "external"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order_id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    }
                }
            }
        },
        "/v1/order/{order_id}/status": {
            "get": {
                "tags": [
                    "external"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order_id",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.OrderStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_errors.ErrorResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_kwaaka-team_orders-core_core_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "description": {
                    "type": "string"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kwaaka-team_orders-core_core_externalapi_models.CreationResult": {
            "type": "object",
            "properties": {
                "orderId": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "github_com_kwaaka-team_orders-core_core_externalapi_models.DeliveryInfo": {
            "type": "object",
            "properties": {
                "clientArrivementDate": {
                    "type": "string"
                },
                "clientName": {
                    "type": "string"
                },
                "courierArrivementDate": {
                    "type": "string"
                },
                "deliveryAddress": {
                    "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.MarketPlaceDeliveryAddress"
                },
                "deliveryDate": {
                    "description": "CourierArrivementDate string       `json:\"courierArrivementDate\"`",
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kwaaka-team_orders-core_core_externalapi_models.MarketPlaceDeliveryAddress": {
            "type": "object",
            "properties": {
                "full": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                }
            }
        },
...
        "github_com_kwaaka-team_orders-core_core_externalapi_models.Order": {
            "description": "Discriminator Дискриминатор схемы обьекта. Для MarketplaceOrder равен EatsId Сквозной идентификатор заказа на стороне Яндекс.Еды в формате DDDDDD-DDDDDD",
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "deliveryInfo": {
                    "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.DeliveryInfo"
                },
                "discriminator": {
                    "type": "string",
                    "example": "marketplace"
                },
                "eatsId": {
                    "type": "string",
                    "example": "190330-123456"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.OrderItem"
                    }
                },
                "paymentInfo": {
                    "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.PaymentInfo"
                },
                "persons": {
                    "type": "integer"
                },
                "promos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.OrderPromo"
                    }
                },
                "restaurantId": {
                    "type": "string"
                }
            }
        },
        "github_com_kwaaka-team_orders-core_core_externalapi_models.OrderItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "modifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.OrderModification"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "promos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kwaaka-team_orders-core_core_externalapi_models.OrderPromo"
                    }
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
        "github_com_kwaaka-team_orders-core_core_externalapi_models.OrderModification": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "github_com_kwaaka-team_orders-core_core_externalapi_models.OrderPromo": {
            "type": "object",
            "properties": {
                "discount": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_kwaaka-team_orders-core_core_externalapi_models.OrderStatusResponse": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kwaaka-team_orders-core_core_externalapi_models.PaymentInfo": {
            "type": "object",
            "properties": {
                "itemsCost": {
                    "type": "integer"
                },
                "paymentType": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Token from cognito\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}